Move types-store to types module
